refactor: 관리자 페이지 UI 분리
